requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
iiif-prezi3==1.1.1
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
pydantic==1.10.8
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.30.0
six==1.16.0
typing_extensions==4.6.1
tzdata==2023.3
urllib3==2.0.2
